Improve code formatting

+++ sample_app/users/models.py
@@ -4,10 +4,11 @@ from sample_app import db, bc
 from utils.data import Model
 
 
-user_authority_table = db.Table("jhi_user_authority",
-                                db.Column("user_id", db.BigInteger, db.ForeignKey('jhi_user.id'), nullable=False),
-                                db.Column("authority_name", db.String(50), db.ForeignKey('jhi_authority.name'), nullable=False),
-                                db.PrimaryKeyConstraint('user_id', 'authority_name'))
+user_authority_table = db.Table(
+    "jhi_user_authority",
+    db.Column("user_id", db.Integer, db.ForeignKey('jhi_user.id'), nullable=False),
+    db.Column("authority_name", db.String(50), db.ForeignKey('jhi_authority.name'), nullable=False),
+    db.PrimaryKeyConstraint('user_id', 'authority_name'))
 
 
 class User(Model):
